Fixed circular depenedency
